use /etc/init.d/ path for unicorn control instead of service
capistrano-unicorn-nginx assumes a server-side user with full sudo access
which is not a secure way to run a webapp. While it would be possible
to limit sudo access to just the 'service' command, this would still
allow the webapp user to shutdown many critical services. Switching
to the direct /etc/init.d/ command allows us to write a specific sudo
grant without giving away the farm.